feat: Switching from the BACnet library to node-backstack. Added support for Multistate objects
Switching from the BACnet library to node-backstack.
When querying for the list of objects (OBJECT_LIST) and there is a large number of objects on the controllers (around 100), readProperty returns an incorrect value. Replaced it with the node-bacstack library.
https://github.com/fh1ch/node-bacstack

Changing from readProperty to readPropertyMultiple when we querying for name of objects.
The change was made to reduce the number of packets when querying for a large number of objects.

Added support for Multistate objects.
Added Multistate Variable, Multistate Input, Multistate Output.